chore: create an applyNewLayerSource() utils function
