+++ src/features/mcda/components/MCDALayerEditor/index.tsx
@@ -1,10 +1,7 @@
 import { useAtom } from '@reatom/npm-react';
 import { useCallback } from 'react';
 import { layersSourcesAtom } from '~core/logical_layers/atoms/layersSources';
-import { store } from '~core/store/store';
-import { createUpdateLayerActions } from '~core/logical_layers/utils/createUpdateActions';
-import { deepCopy } from '~core/logical_layers/utils/deepCopy';
-import { enabledLayersAtom } from '~core/logical_layers/atoms/enabledLayers';
+import { applyNewMCDAConfig } from '~features/mcda';
 import s from './style.module.css';
 import { MCDALayerDetails } from './MCDALayerDetails/MCDALayerDetails';
 import type {
@@ -12,30 +9,6 @@ import type {
   MCDALayer,
 } from '~core/logical_layers/renderers/stylesConfigs/mcda/types';
 import type { LayerEditorProps } from '~core/logical_layers/types/editors';
-import type { Action } from '@reatom/core-v2';
-
-function updateMCDAConfig(config: MCDAConfig) {
-  const id = config.id;
-  const oldSource = layersSourcesAtom.getState().get(id)?.data;
-  if (oldSource) {
-    const newSource = deepCopy(oldSource);
-    if (newSource?.style?.config) {
-      newSource.style.config = { ...config };
-    }
-    const actions: Array<Action> = [
-      enabledLayersAtom.delete(id),
-      ...createUpdateLayerActions([
-        {
-          id,
-          source: newSource,
-        },
-      ]).flat(),
-    ];
-
-    store.dispatch(actions);
-    store.dispatch(enabledLayersAtom.set(id));
-  }
-}
 
 export function MCDALayerEditor({ layerId }: LayerEditorProps) {
   const [mcdaConfig] = useAtom(
@@ -53,7 +26,7 @@ export function MCDALayerEditor({ layerId }: LayerEditorProps) {
           oldLayer.id === editedMCDALayer.id ? editedMCDALayer : oldLayer,
         );
         const editedConfig = { ...mcdaConfig, layers: newLayers };
-        updateMCDAConfig(editedConfig);
+        applyNewMCDAConfig(editedConfig);
       }
     },
     [mcdaConfig],
